Move the code that creates the image to pkg/dmlegacy
